+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixed debug output in readGP
 * Fixed import bug in readGP
+* Fixed bug in addWeightsCorrelated where occasionally the output looked as if -rho was passed
 
 # grapherator 1.0.0